requirements.txt
pyyaml
toml
fedora_messaging
pygit2
pygithub
aiofiles
aiohttp
tox
pytest
pytest-asyncio